requirements.txt
cairosvg>=2.5.0
numpy>=1.20.0
Pillow>=8.0.0
scipy>=1.12
fastapi[standard]>=0.115
uvicorn[standard]>=0.34
python-multipart>=0.0.18
